chore(react-tree): adds Adam's variant C example as reference
