+++ packages/react-tree-grid/stories/AdamsTreeGridActiveOnlyEntireRowNarration.stories.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react';
+import { AccessibleMeetBase } from './Adam/AccessibleMeetBase';
+
+export const AdamsTreeGridActiveOnlyEntireRowNarration = () => (
+  <AccessibleMeetBase variant="gridsActiveOnlyEntireRowNarration" />
+);
