--- packages/react-tree-grid/stories/VariantB/index.ts
@@ -1,1 +1,0 @@
-export * from './TreeGridWithEnterInputsRenderer';
