+++ packages/react-tree-grid/stories/index.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta } from '@storybook/react';
 import { TreeGrid } from '@fluentui-contrib/react-tree-grid';
 
-export { VariantB } from './TreeGridVariantB.stories';
+export { AdamsTreeGridActiveOnlyEntireRowNarration } from './AdamsTreeGridActiveOnlyEntireRowNarration.stories';
 
 const meta: Meta<typeof TreeGrid> = {
   title: 'TreeGrid',
